requirements.txt
torch
torchvision==0.2.2
numpy
pillow
scikit-learn
tqdm
matplotlib
visdom
